make a unique inference engine class
