+++ tools/run_webui.py
@@ -11,9 +11,9 @@ pyrootutils.setup_root(__file__, indicator=".project-root", pythonpath=True)
 from tools.llama.generate import launch_thread_safe_queue
 from tools.schema import ServeTTSRequest
 from tools.vqgan.inference import load_model as load_decoder_model
-from tools.webui.inference_engine import InferenceEngine
-from tools.webui.inference_engine.utils import get_inference_wrapper
-from tools.webui.web_app import build_app
+from tools.inference_engine import InferenceEngine
+from tools.webui.inference import get_inference_wrapper
+from tools.webui import build_app
 
 # Make einx happy
 os.environ["EINX_FILTER_TRACEBACK"] = "false"
